Renamed base_command.rb, fixed indentation and fixed variable naming inconsistency
