+++ metalware-new/src/commands/base_command.rb
