+++ metalware-new/src/commands/dhcp.rb
@@ -1,5 +1,5 @@
 
-require 'commands/basecommand'
+require 'commands/base_command'
 require 'constants'
 require 'output'
 require 'templater'
