+++ metalware-new/src/commands/hosts.rb
@@ -1,5 +1,5 @@
 
-require 'commands/basecommand'
+require 'commands/base_command'
 require 'constants'
 require 'iterator'
 require 'nodes'
